Revert "refactor: 운동거리 int로 변경"
This reverts commit 47e762dae46edaafe2e00d12889fbad1324a03db.

+++ src/main/java/clofi/runningplanet/running/dto/RunningStatusResponse.java
@@ -7,7 +7,7 @@ public record RunningStatusResponse(
 	Long memberId,
 	String nickname,
 	int runTime,
-	int runDistance,
+	double runDistance,
 	boolean isEnd
 ) {
 	public RunningStatusResponse(Member member, Record savedRecord) {
